Update README.md
Coursera course: HTML, CSS, and Javascript for Web Developers; Module 4 Assignment. We use Javascript in order to loop over an array of names and print out either hello or goodbye to that name to the browser console. If the name starts with the letter j or J, you are to print out Goodbye JSomeName. If the name starts with any other letter, you are to print out Hello SomeName.